Disabled doctitle_xform for rest rendering
Restructured text was previously being rendered so that:

	Header1
	=======

	Some text.

	Header2
	-------

	Some text.

	Header3
	~~~~~~~

	Some more text.

Had a tag pattern of:

	<h1 class="title">Header1</h1>
	...
	<h1>Header2</h1>
	...
	<h2>Header3</h2>

Which does not suit the github css stylesheets well, as the
class="title" is ignored so the first two headings are rendered the same
size. Additionally the github stylesheets add a horizontal rule/line
above <h2> headings which in rst are really level 3 headings in the
source document.

Adding doctitle_xform: False to the render settings results in:

	<h1>Header1</h1>
	...
	<h2>Header2</h2>
	...
	<h3>Header3</h3>

Which suits the github stylesheets much better and matches the markdown
results.

The tests have also been updated to include appropriate headings in the
rst example.