Burn `require 'markdown'` with the fire of a thousand suns
GitHub::Markup now has an actual preference order for Markdown libraries,
tries to require them explicitly and use their actual APIs instead of
the compatibility 'Markdown' layer.

Redcarpet is the default library, and the one listed in the Gemfile.
If it's not available, the (opinionated) require order is:

	- Redcarpet
	- RDiscount
	- Maruku
	- Kramdown
	- BlueCloth

This soothes my soul.