Use CC0 as the license for rst2html.
This should clarify the kerfuffle merged in #177 and is in spirit of
what I intended when I wrote the file in 2008. See the CC0 FAQ for more
details: http://creativecommons.org/about/cc0

I used the git log to mention the people who contributed to the file
since then.

Fixes #205.