Make rest2html program run more easily
As interpreted by a Python 2.7 interpreter on a GNU system.

This shebang form is unusual (env python2 -S), nor is relying on
whitespace between shebang entries considered portable, de-facto nor
de-jure.

In addition, reloading sys seems to be necessary to allow for
setdefaultencoding.  That's apparently to prevent people from using
that symbol, but as it was working before somehow, just fix it.