Interpret titles and subtitles in rst
This hopes to make subtitles usable without changing the markup and
table-of-contents generation of documents much at all.

Previously, titles and subtitles[0] would introduce a lot of nesting
in a generated table of contents.  Enable both doctitle_xform and
sectsubtitle_xform to remove those.

However, merely enabling those will cause the existing top-level,
non-title HTML header sizes to be promoted to <h1>, since titles and
subtitles are not considered "sections", changing how the HTML output
looks a rather lot.

That's not how some sampled rsts (without subtitles) on work today:
the title is biggest at <h1>, and all other sections are <h2> or
below.  So set the initial_header_level to "2", so that the largest
text sections can have is <h2>, tied in size for the subtitle.

[0]: http://docutils.sourceforge.net/docs/user/rst/quickstart.html#document-title-subtitle