rst2html: Extend the field_name_limit to 50
Field lists are rendered weirdly in reST if the field name is longer
that 14 characters. In this case the table cell where the field name is
gets a colspan=2.

For example this:

```rest
:123456789012345: too long, colspan=2
:12345678901234: OK
```

Gets translated to:

```html
<table class="docinfo" frame="void" rules="none">
<col class="docinfo-name" />
<col class="docinfo-content" />
<tbody valign="top">
<tr class="field"><th class="docinfo-name"
colspan="2">123456789012345:</th></tr>
<tr class="field"><td>&nbsp;</td><td class="field-body">too long,
colspan=2</td>
</tr>
<tr class="field"><th class="docinfo-name">12345678901234:</th><td
class="field-body">OK</td>
</tr>
</tbody>
</table>
```

Increasing this default value of 14 to something more reasonable, like
50, seems sensible. Even when this is just another arbitrary value,
since GitHub have a fixed total width, it seems sensible to not make it
much more longer, since if the field name is too long it will look weird
anyway.