Display unknown restructuredText directives
This addresses issue #514.

Previous behavior was to silently ignore unknown directives and
comments, which made .rst documents that utilized rst extensions
appear to be incomplete.  This fix allows all the information to
be displayed, and further allows comments to be placed to indicate
that the github rst viewer is not the preferred viewer for the
document.

This patch will not alter the appearance of any document that
is specifically coded for display at github which doesn't use
restucturedText comments.

Additional possible enhancements include figuring out how to
gray out this text, and/or to support a special github directive
at the start of the text file that controls whether this new
behavior is enabled.