Added doctest blocks to reStructuredText
Many Python projects use Sphinx to create documentation from .rst markup.

Using a doctest:: directive the markup can include blocks of code that is both
tested and rendered in the final documentation. These blocks are not rendered
on github.com though, e.g. in a README.rst.

To remedy this (without adding Sphinx as a dependency) this PR aliases
doctest:: directives to code:: directives,. The language of such blocks is set
to 'python', since doctest only supports Python.

Refs dateutil/dateutil#83